requirements.txt
aioredis==1.3.1
APScheduler==3.9.1
channels==3.0.4
channels-redis==3.4.0
Django==3.2.5
django-cors-headers==3.7.0
django-rest-knox==4.2.0
django-summernote==0.8.20.0
djangochannelsrestframework==1.0.0
djangorestframework==3.12.4
Pillow==8.4.0
pytest==7.0.1
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
psycopg2==2.9.5
gunicorn==21.2.0
whitenoise==6.5.0
